+++ Calculadora2JEE-ejb/src/java/co/edu/unipiloto/metodos/session/CalculadoraBeanLocal.java
@@ -14,13 +14,13 @@ import javax.ejb.Local;
 @Local
 public interface CalculadoraBeanLocal {
 
-    Integer suma(int val1, int val2);
+    Double suma(double val1, double val2);
 
-    Integer resta(int val1, int val2);
+    Double resta(double val1, double val2);
 
-    Integer multiplicacion(int val1, int val2);
+    Double multiplicacion(double val1, double val2);
 
-    Integer modulo(int val1, int val2);
+    Double modulo(double val1, double val2);
 
     Double division(double val1, double val2);
